Struct srml_session::Module

pub struct Module<T: Trait>(_);

Methods

impl<T: 'static + Trait> Module<T>

pub fn validators() -> Vec<T::AccountId>

The current set of validators.

pub fn length() -> T::BlockNumber

Current length of the session.

pub fn current_index() -> T::BlockNumber

Current index of the session.

pub fn current_start() -> T::Moment

Timestamp when current session started.

pub fn forcing_new_session() -> Option<bool>

New session is being forced if this entry exists; in which case, the boolean value is whether the new session should be considered a normal rotation (rewardable) or exceptional (slashable).

impl<T: Trait> Module<T>

pub fn validator_count() -> u32

The current number of validators.

pub fn last_length_change() -> T::BlockNumber

The last length change if there was one, zero if not.

pub fn apply_force_new_session(apply_rewards: bool) -> Result

Forces a new session, no origin.

pub fn set_validators(new: &[T::AccountId])

Set the current set of validators.

Called by staking::new_era only. rotate_session must be called after this in order to update the session keys to the next validator set.

pub fn check_rotate_session(block_number: T::BlockNumber)

Hook to be called after transaction processing.

pub fn rotate_session(is_final_block: bool, apply_rewards: bool)

Move on to next session: register the new authority set.

pub fn ideal_session_duration() -> T::Moment

Get the time that should have elapsed over a session if everything was working perfectly.

pub fn blocks_remaining() -> T::BlockNumber

Number of blocks remaining in this session, not counting this one. If the session is due to rotate at the end of this block, then it will return 0. If the session just began, then it will return Self::length() - 1.