Trait srml_session::OnSessionEnding

pub trait OnSessionEnding<ValidatorId> {
    fn on_session_ending(
        ending_index: SessionIndex, 
        will_apply_at: SessionIndex
    ) -> Option<Vec<ValidatorId>>;
}

An event handler for when the session is ending. TODO [slashing] consider renaming to OnSessionStarting

Required methods

fn on_session_ending(
ending_index: SessionIndex,
will_apply_at: SessionIndex
) -> Option<Vec<ValidatorId>>

Handle the fact that the session is ending, and optionally provide the new validator set.

Even if the validator-set is the same as before, if any underlying economic conditions have changed (i.e. stake-weights), the new validator set must be returned. This is necessary for consensus engines making use of the session module to issue a validator-set change so misbehavior can be provably associated with the new economic conditions as opposed to the old.

ending_index is the index of the currently ending session. The returned validator set, if any, will not be applied until will_apply_at. will_apply_at is guaranteed to be at least ending_index + 1, since session indices don't repeat, but it could be some time after in case we are staging authority set changes.

Loading content...