Trait srml_session::SessionHandler

pub trait SessionHandler<ValidatorId> {
    fn on_genesis_session<Ks: OpaqueKeys>(validators: &[(ValidatorId, Ks)]);

    fn on_new_session<Ks: OpaqueKeys>(
        changed: bool, 
        validators: &[(ValidatorId, Ks)], 
        queued_validators: &[(ValidatorId, Ks)]
    );

    fn on_disabled(validator_index: usize);

    fn on_before_session_ending() { ... }
}

Handler for session lifecycle events.

Required methods

fn on_genesis_session<Ks: OpaqueKeys>(validators: &[(ValidatorId, Ks)])

The given validator set will be used for the genesis session. It is guaranteed that the given validator set will also be used for the second session, therefore the first call to on_new_session should provide the same validator set.

fn on_new_session<Ks: OpaqueKeys>(
changed: bool,
validators: &[(ValidatorId, Ks)],
queued_validators: &[(ValidatorId, Ks)]
)

Session set has changed; act appropriately. Note that this can be called before initialization of your module.

changed is true whenever any of the session keys or underlying economic identities or weightings behind those keys has changed.

fn on_disabled(validator_index: usize)

A validator got disabled. Act accordingly until a new session begins.

Loading content...
Provided methods

fn on_before_session_ending()

A notification for end of the session.

Note it is triggered before any OnSessionEnding handlers, so we can still affect the validator set.

Loading content...