Crate srml_session

Session manager: is told the validators and allows them to manage their session keys for the consensus module.

Structs

CurrentIndex
Current index of the session.

CurrentStart
Timestamp when current session started.

ForcingNewSession
New session is being forced if this entry exists; in which case, the boolean value is whether the new session should be considered a normal rotation (rewardable) or exceptional (slashable).

GenesisConfig

Module

SessionLength
Current length of the session.

Validators
The current set of validators.

Enums

Call

RawEvent
Events for this module.

Traits

OnSessionChange
A session has changed.

Trait

Type Definitions

Event
RawEvent specialized for the configuration Trait